% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Calc_IS_score.R
\name{calculate_im_score}
\alias{calculate_im_score}
\title{Calculate Immunomodulatory Score}
\usage{
calculate_im_score(otu_table, binding_score, w_P = 0.3, w_A = 0.3, w_I = 0.4)
}
\arguments{
\item{otu_table}{A matrix or data frame of relative abundances (species x samples).}

\item{binding_score}{A data frame containing inferred binding scores either using Kau Index, IgA probability score or IgA probability ratio (species x samples).}

\item{w_P}{Weight for prevalence in the IM score (0.3).}

\item{w_A}{Weight for relative abundance in the IM score (0.3).}

\item{w_I}{Weight for the binding score in the IM score (0.4).}
}
\value{
A data frame with Species, Sample, Binding_score, Prevalence, Rel_ab, and IM_score.
}
\description{
This function merges relative abundance (`otu_table`), prevalence, and Ig binding score data (`binding_score`)
to compute an Immunomodulatory Score (IM_score) for each OTU (species).
}
\examples{
# result <- calculate_im_score(otu_table, binding_score)
}
